requirements.txt
langchain==0.1.13 
langsmith==0.1.31
langchain-community==0.0.29
langchain-openai==0.1.1
langchain-google-genai==1.0.1
langchain-anthropic==0.1.4
langchain-fireworks==0.1.1
langserve==0.0.51
fastapi==0.110.0
pydantic==2.6.4
uvicorn==0.23.2
beautifulsoup4==4.12.3
psycopg2==2.9.9
lxml==5.2.1
voyageai==0.2.1
pillow==10.2.0
tiktoken==0.6.0
cohere==5.2.5
gradio==4.22.0 
gradio-client==0.13.0
weaviate-client==4.5.5
parse==1.20.2